ai-revenue-copilot/requirements.txt
fastapi
uvicorn
groq
pydantic
pydantic-settings
python-dotenv
